lift win32 dialogs into separate files
