+++ src/Plugins.TASVideo/Common.hpp
@@ -9,10 +9,7 @@
 #include <CommonPCH.hpp>
 #include <VersionNameHelpers.hpp>
 #include <m64rr/API.hpp>
-#include <windows.h>
 #include <GL/glew.h>
-#include <commctrl.h>
-#include <windowsx.h>
 #include <emmintrin.h>
 #include <intrin.h>
 #include <xbrz.h>
@@ -20,9 +17,6 @@
 #include <GL/glext.h>
 #include <GL/wglext.h>
 #include "Types.hpp"
-
-#define min(a, b) ((a) < (b) ? (a) : (b))
-#define max(a, b) ((a) > (b) ? (a) : (b))
 
 #define DEBUG_ERROR 0
 #define DEBUG_LOW 0
