+++ src/Plugins.TASVideo/Config.cpp
@@ -7,67 +7,13 @@
 #include "Common.hpp"
 #include "Config.hpp"
 #include "glN64.hpp"
-#include "resource.h"
 #include "RSP.hpp"
 #include "Textures.h"
 #include "OpenGL.hpp"
-
-#undef min
-#undef max
 #include <nlohmann/json.hpp>
-#define min(a, b) ((a) < (b) ? (a) : (b))
-#define max(a, b) ((a) > (b) ? (a) : (b))
-
-HWND hConfigDlg;
-
-#define numWindowedModes 12
-#define CONFIG_FILE_NAME "TASVideo.json"
-
 using nlohmann::json;
 
-struct ResolutionPreset
-{
-    uint32_t width{};
-    uint32_t height{};
-    std::string description;
-};
-
-const std::vector<ResolutionPreset> RESOLUTION_PRESETS = {
-    {320, 240, "320 x 240 (4:3)"},     {400, 300, "400 x 300 (4:3)"},      {480, 360, "480 x 360 (4:3)"},
-    {640, 480, "640 x 480 (4:3)"},     {800, 600, "800 x 600 (4:3)"},      {960, 720, "960 x 720 (4:3)"},
-    {1024, 768, "1024 x 768 (4:3)"},   {1152, 864, "1152 x 864 (4:3)"},    {1280, 720, "1280 x 720 (16:9)"},
-    {1280, 960, "1280 x 960 (4:3)"},   {1280, 1024, "1280 x 1024 (5:4)"},  {1440, 1080, "1440 x 1080 (4:3)"},
-    {1600, 1200, "1600 x 1200 (4:3)"}, {1920, 1080, "1920 x 1080 (16:9)"}, {2560, 1440, "2560 x 1440 (16:9)"},
-    {3840, 2160, "3840 x 2160 (16:9)"}};
-
-const std::vector<std::pair<uint8_t, std::string>> FILTER_NAMES = {
-    {0, "Default"},
-    {1, "Always Smooth"},
-    {2, "Always Pixelated"},
-};
-
-const std::vector<std::string> ASPECT_MODE_NAMES = {
-    "Pillarbox",
-    "Stretch",
-    "Widescreen",
-};
-
-constexpr AspectMode DEFAULT_ASPECT_MODE = AspectMode::Widescreen;
-
-static AspectMode to_aspect_mode(int value)
-{
-    if (value < 0 || value >= (int)ASPECT_MODE_NAMES.size()) return DEFAULT_ASPECT_MODE;
-    return (AspectMode)value;
-}
-
-static std::optional<ResolutionPreset> get_preset_by_resolution(uint32_t width, uint32_t height)
-{
-    for (const auto &preset : RESOLUTION_PRESETS)
-    {
-        if (preset.width == width && preset.height == height) return preset;
-    }
-    return std::nullopt;
-}
+#define CONFIG_FILE_NAME "TASVideo.json"
 
 static std::filesystem::path get_config_path()
 {
@@ -146,210 +92,4 @@ void Config_SaveConfig()
 
     std::ofstream ofs(get_config_path());
     ofs << std::setw(2) << j;
-}
-
-void Config_ApplyDlgConfig(HWND hWndDlg)
-{
-    static GLInfo prev_OGL{};
-
-    prev_OGL = OGL;
-
-    char text[256]{};
-    int i;
-
-    Edit_GetText(GetDlgItem(hWndDlg, IDC_CACHEMEGS), text, 4);
-    cache.maxBytes = atol(text) * 1048576;
-
-    OGL.textureFilter = (TextureFilter)SendDlgItemMessage(hWndDlg, IDC_TEXTUREFILTER, CB_GETCURSEL, 0, 0);
-    OGL.filterScale = SendDlgItemMessage(hWndDlg, IDC_FSCALE, TBM_GETPOS, 0, 0);
-    OGL.fog = (SendDlgItemMessage(hWndDlg, IDC_FOG, BM_GETCHECK, 0, 0) == BST_CHECKED);
-    OGL.msaa = (SendDlgItemMessage(hWndDlg, IDC_MSAA, BM_GETCHECK, 0, 0) == BST_CHECKED) ? 4 : 0;
-    OGL.originAdjust = (OGL.textureFilter == TextureFilter::SaI ? 0.25 : 0.50);
-    OGL.ignoreScissor = (SendDlgItemMessage(hWndDlg, IDC_SCISSOR, BM_GETCHECK, 0, 0) == BST_CHECKED);
-    OGL.clear_override = (SendDlgItemMessage(hWndDlg, IDC_CLEAR, BM_GETCHECK, 0, 0) == BST_CHECKED);
-
-    char val[32]{};
-    SendMessage(GetDlgItem(hWndDlg, IDC_WINDOWED_X), WM_GETTEXT, std::size(val), (LPARAM)val);
-    OGL.windowedWidth = atoi(val);
-    SendMessage(GetDlgItem(hWndDlg, IDC_WINDOWED_Y), WM_GETTEXT, std::size(val), (LPARAM)val);
-    OGL.windowedHeight = atoi(val);
-
-    OGL.smoothing = ComboBox_GetCurSel(GetDlgItem(hWndDlg, IDC_SMOOTHING));
-    OGL.aspectMode = to_aspect_mode(ComboBox_GetCurSel(GetDlgItem(hWndDlg, IDC_ASPECT)));
-
-    OGL.usePolygonStipple = (SendDlgItemMessage(hWndDlg, IDC_DITHEREDALPHATEST, BM_GETCHECK, 0, 0) == BST_CHECKED);
-
-    Config_SaveConfig();
-    Config_LoadConfig();
-
-    const auto needs_restart =
-        OGL.smoothing != prev_OGL.smoothing || OGL.textureFilter != prev_OGL.textureFilter ||
-        OGL.filterScale != prev_OGL.filterScale || OGL.msaa != prev_OGL.msaa ||
-        OGL.ignoreScissor != prev_OGL.ignoreScissor || OGL.clear_override != prev_OGL.clear_override ||
-        OGL.windowedWidth != prev_OGL.windowedWidth || OGL.windowedHeight != prev_OGL.windowedHeight ||
-        OGL.usePolygonStipple != prev_OGL.usePolygonStipple;
-
-    if (RSP.thread && needs_restart)
-    {
-        SetEvent(RSP.threadMsg[RSPMSG_RESTART]);
-        WaitForSingleObject(RSP.threadFinished, INFINITE);
-    }
-}
-
-static void select_resolution_in_combobox(HWND cb_hwnd, uint32_t width, uint32_t height)
-{
-    for (size_t i = 0; i < RESOLUTION_PRESETS.size(); i++)
-    {
-        const auto &preset = RESOLUTION_PRESETS[i];
-        if (width == preset.width && height == preset.height)
-        {
-            ComboBox_SetCurSel(cb_hwnd, i);
-            return;
-        }
-    }
-
-    ComboBox_SetCurSel(cb_hwnd, -1);
-}
-
-BOOL CALLBACK ConfigDlgProc(HWND hWndDlg, UINT message, WPARAM wParam, LPARAM lParam)
-{
-    switch (message)
-    {
-    case WM_INITDIALOG: {
-        hConfigDlg = hWndDlg;
-
-        const auto cb_hwnd = GetDlgItem(hWndDlg, IDC_WINDOWEDRES);
-
-        for (const auto &preset : RESOLUTION_PRESETS)
-        {
-            ComboBox_AddString(cb_hwnd, preset.description.c_str());
-        }
-
-        for (const auto &filter : FILTER_NAMES)
-        {
-            ComboBox_AddString(GetDlgItem(hWndDlg, IDC_SMOOTHING), filter.second.c_str());
-        }
-        ComboBox_SetCurSel(GetDlgItem(hWndDlg, IDC_SMOOTHING), (int)OGL.smoothing);
-
-        for (const auto &aspect_mode : ASPECT_MODE_NAMES)
-        {
-            ComboBox_AddString(GetDlgItem(hWndDlg, IDC_ASPECT), aspect_mode.c_str());
-        }
-        ComboBox_SetCurSel(GetDlgItem(hWndDlg, IDC_ASPECT), (int)OGL.aspectMode);
-
-        select_resolution_in_combobox(cb_hwnd, OGL.windowedWidth, OGL.windowedHeight);
-
-        SendDlgItemMessage(hWndDlg, IDC_WINDOWED_X, WM_SETTEXT, 0, (LPARAM)std::to_string(OGL.windowedWidth).c_str());
-        SendDlgItemMessage(hWndDlg, IDC_WINDOWED_Y, WM_SETTEXT, 0, (LPARAM)std::to_string(OGL.windowedHeight).c_str());
-
-        SendDlgItemMessage(hWndDlg, IDC_TEXTUREFILTER, CB_ADDSTRING, 0, (LPARAM) "None");
-        SendDlgItemMessage(hWndDlg, IDC_TEXTUREFILTER, CB_ADDSTRING, 0, (LPARAM) "2xSaI");
-        SendDlgItemMessage(hWndDlg, IDC_TEXTUREFILTER, CB_ADDSTRING, 0, (LPARAM) "xBRZ");
-        SendDlgItemMessage(hWndDlg, IDC_TEXTUREFILTER, CB_ADDSTRING, 0, (LPARAM) "Hqx");
-        SendDlgItemMessage(hWndDlg, IDC_TEXTUREFILTER, CB_SETCURSEL, (int)OGL.textureFilter, 0);
-        SendMessage(GetDlgItem(hWndDlg, IDC_FSCALE), TBM_SETPOS, TRUE, OGL.filterScale);
-
-        SendDlgItemMessage(hWndDlg, IDC_SCISSOR, BM_SETCHECK,
-                           OGL.ignoreScissor ? (LPARAM)BST_CHECKED : (LPARAM)BST_UNCHECKED, 0);
-        SendDlgItemMessage(hWndDlg, IDC_CLEAR, BM_SETCHECK,
-                           OGL.clear_override ? (LPARAM)BST_CHECKED : (LPARAM)BST_UNCHECKED, 0);
-
-        // Enable/disable fog
-        SendDlgItemMessage(hWndDlg, IDC_FOG, BM_SETCHECK, OGL.fog ? (LPARAM)BST_CHECKED : (LPARAM)BST_UNCHECKED, 0);
-        SendDlgItemMessage(hWndDlg, IDC_MSAA, BM_SETCHECK, OGL.msaa == 4 ? (LPARAM)BST_CHECKED : (LPARAM)BST_UNCHECKED,
-                           0);
-
-        SendDlgItemMessage(hWndDlg, IDC_DITHEREDALPHATEST, BM_SETCHECK,
-                           OGL.usePolygonStipple ? (LPARAM)BST_CHECKED : (LPARAM)BST_UNCHECKED, 0);
-
-        const auto cache_size = std::to_string(cache.maxBytes / 1048576);
-        SendDlgItemMessage(hWndDlg, IDC_CACHEMEGS, WM_SETTEXT, 0, (LPARAM)cache_size.c_str());
-
-        SendMessage(hWndDlg, WM_COMMAND, MAKEWPARAM(IDC_TEXTUREFILTER, CBN_SELCHANGE),
-                    (LPARAM)GetDlgItem(hWndDlg, IDC_TEXTUREFILTER));
-
-        return TRUE;
-    }
-    case WM_COMMAND:
-        switch (LOWORD(wParam))
-        {
-        case IDOK:
-            Config_ApplyDlgConfig(hWndDlg);
-            EndDialog(hWndDlg, wParam);
-            hConfigDlg = NULL;
-            return TRUE;
-        case IDCANCEL:
-            EndDialog(hWndDlg, wParam);
-            hConfigDlg = NULL;
-            return TRUE;
-        case IDC_WINDOWEDRES:
-            if (HIWORD(wParam) == CBN_SELCHANGE)
-            {
-                const auto cb_hwnd = (HWND)lParam;
-                const auto index = ComboBox_GetCurSel(cb_hwnd);
-
-                const auto &preset = RESOLUTION_PRESETS[index];
-                Edit_SetText(GetDlgItem(hWndDlg, IDC_WINDOWED_X), std::to_string(preset.width).c_str());
-                Edit_SetText(GetDlgItem(hWndDlg, IDC_WINDOWED_Y), std::to_string(preset.height).c_str());
-            }
-            break;
-        case IDC_WINDOWED_X:
-        case IDC_WINDOWED_Y:
-            if (HIWORD(wParam) == EN_CHANGE)
-            {
-                char w_str[32]{};
-                char h_str[32]{};
-                Edit_GetText(GetDlgItem(hWndDlg, IDC_WINDOWED_X), w_str, std::size(w_str));
-                Edit_GetText(GetDlgItem(hWndDlg, IDC_WINDOWED_Y), h_str, std::size(h_str));
-
-                uint32_t width{};
-                uint32_t height{};
-                try
-                {
-                    width = std::stoul(w_str);
-                    height = std::stoul(h_str);
-                }
-                catch (...)
-                {
-                    break;
-                }
-
-                select_resolution_in_combobox(GetDlgItem(hWndDlg, IDC_WINDOWEDRES), width, height);
-            }
-            break;
-        case IDC_TEXTUREFILTER:
-            if (HIWORD(wParam) == CBN_SELCHANGE)
-            {
-                const auto filter = (TextureFilter)SendDlgItemMessage(hWndDlg, IDC_TEXTUREFILTER, CB_GETCURSEL, 0, 0);
-
-                int min_scale = 0;
-                int max_scale = 0;
-                switch (filter)
-                {
-                case TextureFilter::SaI:
-                    min_scale = max_scale = 2;
-                    break;
-                case TextureFilter::xBRZ:
-                    min_scale = 2;
-                    max_scale = 6;
-                    break;
-                case TextureFilter::Hqx:
-                    min_scale = 2;
-                    max_scale = 4;
-                    break;
-                default:
-                    break;
-                }
-
-                SendMessage(GetDlgItem(hWndDlg, IDC_FSCALE), TBM_SETRANGE, TRUE, MAKELONG(min_scale, max_scale));
-                EnableWindow(GetDlgItem(hWndDlg, IDC_FSCALE), min_scale != max_scale);
-            }
-        }
-    }
-    return FALSE;
-}
-
-void Config_Show(HWND parent)
-{
-    DialogBox(g_tas_ctx.hinst, MAKEINTRESOURCE(IDD_CONFIGDLG), parent, (DLGPROC)ConfigDlgProc);
 }
