+++ src/Plugins.TASVideo/OpenGL.cpp
@@ -51,7 +51,7 @@ void OGL_InitExtensions()
     }
 
     glGetIntegerv(GL_MAX_TEXTURE_IMAGE_UNITS, &OGL.maxTextureUnits);
-    OGL.maxTextureUnits = min(8, OGL.maxTextureUnits); // The plugin only supports 8, and 4 is really enough
+    OGL.maxTextureUnits = std::min(8, OGL.maxTextureUnits); // The plugin only supports 8, and 4 is really enough
 
     OGL.EXT_fog_coord = GLEW_EXT_fog_coord;
     OGL.EXT_secondary_color = GLEW_EXT_secondary_color;
@@ -509,10 +509,10 @@ void OGL_AddTriangle(SPVertex *vertices, int v0, int v1, int v2)
         if ((gSP.geometryMode & G_FOG) && OGL.fog)
         {
             if (vertices[v[i]].z < -vertices[v[i]].w)
-                OGL.vertices[OGL.numVertices].fog = max(0.0f, -(float)gSP.fog.multiplier + (float)gSP.fog.offset);
+                OGL.vertices[OGL.numVertices].fog = std::max(0.0f, -(float)gSP.fog.multiplier + (float)gSP.fog.offset);
             else
                 OGL.vertices[OGL.numVertices].fog =
-                    max(0.0f, vertices[v[i]].z / vertices[v[i]].w * (float)gSP.fog.multiplier + (float)gSP.fog.offset);
+                    std::max(0.0f, vertices[v[i]].z / vertices[v[i]].w * (float)gSP.fog.multiplier + (float)gSP.fog.offset);
         }
 
         if (combiner.usesT0)
