+++ src/Plugins.TASVideo/Resource.h
